Create a new renderer for each new plot type
Currently, the test whether the renderer exists is not sufficient
to decide whether a new renderer needs to be created. This was
causing different plots to use the same renderer.

This change introduces a new argument to the fitToViewport call
that forces it to create a new renderer if requested. This will
enable all the pipeline plot calls to create their own
independent renderer.

Fixes issue #1338