BUG #1739: fitToViewport uses dataset bounds instead of recomputing them
fitToViewport recomputes the dataset bounds using an 2D array of points.
This is slow and imprecise and it does not take into account the bounds stored
in the file. Now we use a new function fitToViewportBounds that
receives the dataset bounds as parameter. The old function is still used for
the outline and the continents.  It is going to be removed in the future.