Generate patterns for transformed actors for fillarea objects
This change makes sure that patterns are generated after calling
`VTKPlots::fitToViewport` on the original dataset. `fitToViewport`
applies a user transform on the actor. Applying that transform on the
input dataset before generating patterns helps remove transform issues.

In this case the management of color arrays is crucial. For solid fill
areas, we create a single polydata with multiple cells and the scalar
array consists of cell colors. For patterns / hatches, we create
multiple polydata per polygon. The original scalar array becomes our
background color (which is transparent white) and each individual
hatch polydata uses the color for that polygon for its own scalar array.