2072 export patterns as vectors (#23)
* Initial work to export patterns as vectors

Disable rasterizing the background before any vector export.

* Fix vector export issue with erroneous triangles

Turns out the clip filter was writing out -inf for all cells outside the clipped
area. Changing the out value to -1 solved this issue.
The size of the patterned polydata should be such that points are all enclosed
in the boundaries of the patterned surface.

* Initial working version of creating patterns via Glyph2D

Each pattern/hatch is a single polydata with standard parameters
that is replicated over each point of the planar pointset using
vtkGlyph2D.

* Added support for 20 different pattern objects

* Fix overlapping patterns issue

The idea is to select a proper out value for the type of dataset
when creating the implicit function used to clip the patterned plane.

* Fix issues with patterns/hatches for boxfill plots

Handle the case where there are no point scalars

* Adjust scale for better support of different aspect ratios

* Better pattern clipping with new vtkCookieCutter

* Generate patterns for transformed actors for isofill plots

This change makes sure that patterns are generated after calling
`VTKPlots::fitToViewport` on the original dataset. `fitToViewport`
applies a user transform on the actor. Applying that transform on the
input dataset before generating patterns helps remove transform issues.

* Generate patterns for transformed actors for fillarea objects

This change makes sure that patterns are generated after calling
`VTKPlots::fitToViewport` on the original dataset. `fitToViewport`
applies a user transform on the actor. Applying that transform on the
input dataset before generating patterns helps remove transform issues.

In this case the management of color arrays is crucial. For solid fill
areas, we create a single polydata with multiple cells and the scalar
array consists of cell colors. For patterns / hatches, we create
multiple polydata per polygon. The original scalar array becomes our
background color (which is transparent white) and each individual
hatch polydata uses the color for that polygon for its own scalar array.

* Choose resolution of pattern polydata based on display scale

This change adds smart determination for resolution of the pattern
polydata based on the rendered screen size. It fixes the issue where the
patterns behave differently on different screens / zooms.

* Fix pattern scaling to render synchronous to screen pixels

This change makes sure that patterns are visible appropriately on all
screen sizes / zooms

* Generate patterns for transformed actors for boxfill plots

This change makes sure that patterns are generated after calling
`VTKPlots::fitToViewport` on the original dataset. `fitToViewport`
applies a user transform on the actor. Applying that transform on the
input dataset before generating patterns helps remove transform issues.

* Avoid adding unrequired actors to fillarea objects

This change fixes the issue where additional transparent actors were
being added to the scene throwing off certain pdf viewers

* Avoid adding the background for patterns/hatches to the renderer

This change makess sure that transparency agnostic PDF viewers would be
able to display patterned areas properly.

* Resolution of patterns and hatches to alleviate aliasing issues

* Fixed the boxfill pipeline to auto-compute scale and resolution

* Compute pattern scale based on the pixel spacing in world coordinates

This change makes the spacing between pattern points to be 10 pixels and
the scale to be 7.5 pixels.

* Added fillareapixelspacing and fillareapixelscale to VCS

These two parameters are active when there is a pattern or hatch being
drawn on the fillarea. The individual pattern elements will be
pixelspacing apart and scaled by pixelscale.

* Defaults for fillareapixelspacing and fillareapixelscale

The values default to 15 pixel spacing per 1000 pixel render window size
and 12 pixel scale respectively.

* Added fillareapixelspacing and fillareapixelscale to meshfill

* Adapt the meshfill pipeline to the new patterning code

* Remove background actors from meshfill patterns

These actors are transparent and not needed in the view

* Respect pattern pixel spacing and scale in color bars

The patterns now look the same in the plot and the color bar

* Ensure that the pixelspacing is always an integer

* WIP: Replace exisiting patterns with custom code

* Changed to better patterns

* Re-create old patterns in the new vector format

Adjusted the code so that scaling is applied appropriately to each
dimension of the pattern.
Auto-computed spacing is always a square now.

* Improve thick diagonal lines and remove debug print

* Compensate for difference in viewport size of the colorbar