Fix issues using Context2D in the presence of axis scaling
Discovered that we were already calling fitToViewport on every dataset
belonging to a Pipeline2D, we were just throwing away the results (or
maybe only using them for picking).

So I have refactored a bit to allow the pipelines to make use of the
results of that initial scaling (without creating/adding renderers,
or adding actors to renderers).  This avoids duplicated computation,
and ensures that the plots come out right when axes are scaled.