Fix plot alignment issues by setting axis margins to zero
This solves the problem with plots and templates being slightly
mis-aligned, as well as problems with the labels seemingly being
cut off on the left and right edges.

With this fix, many of the baselines are now sufficient so that we
will not need to update them.

However, most of the basic plots that did not pass on proj = 0, and
never had a chance to get to proj = -3, are now failing on proj = -3
with a completely blank image.