Support transparent color for logo image
This color only has any effect when exporting plots to pdf.  Due to
the fact that transparency cannot be handled by backend pdf context
device when drawing images, we provide this property to allow user
to pick a transparent color for the logo that matches the background
color of the canvas.  Since the default canvas background color is
white, the default for the logo_transparentcolor is white as well.