Fixes plots shifted down and to the left
By computing screen geometries in floating point, but then truncating
them instead of rounding first, we were inadvertently shifting all
plot components down and to the left.  One effect of this fix is that
we longer see the ~1px gaps between the plot contents and the plot
border on the top and left of the plot.