Add support for the 'reference' attribute on the vector gm (#389)
* do not downgrade

* Add support for the 'reference' attribute on the vector gm

The 'reference' attribute has been available to read/write on
the vector graphics method, but until now has not been implemented.
Now, if it is set, the default behavior of choosing a sensible value
for the arrow length based on the screen space availabe, text size,
etc., will be overridden.  Instead the arrow length will simply be
chosen to match the reference attribute.  Hence, this feature should
be used with care, as it could result in a vector legend with either
very large or very small reference arrow.

* latest vtk and add libcdms in circleci and recipe

* Removed space

* Needs greater equal

* No restriction on vtk

* trying py37 no_coverage needed

* trying to force outputviwer version in testsrunner

* downgrading to py36 since conda bug for testsrunner and py37

* ok pushing back 3.7 now that i uplaod output_viewr for 37

* please no_coverage . merged no_downgrade branch in

removed restriction on libcdms and numpy

* for now forcing 3.6

* trying newewr xcode on mac sidde

* ok no more circleci...

* maybe no comments

* maybe needs two jobs

* forgot the ':' at the end

* templink was left over, that probably was the trick

* ok now with 3.7

* back to 3.6 only for now